+++ sig/herb/engine.rbs
@@ -41,14 +41,6 @@ module Herb
 
     def self.nested_attribute_value: (untyped value) -> untyped
 
-    def self.comment?: (untyped code) -> untyped
-
-    def self.strip_trailing_comment: (untyped code) -> untyped
-
-    def self.prism_available?: () -> untyped
-
-    def self.heredoc?: (untyped code) -> untyped
-
     # : () -> untyped
     def compiler_class: () -> untyped
